admin/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
passlib

python-jose
